refactor(vow-api): sepperate domain and contracts
