+++ vow-api/src/VoW.Api/Contracts/Auth/AuthHandoffRequest.cs
@@ -1,0 +1,6 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace VoW.Api.Contracts.Auth;
+
+public sealed record AuthHandoffRequest(
+    [Required] string Code);
