+++ vow-api/src/VoW.Api/Contracts/Auth/AuthTokenResponse.cs
@@ -1,0 +1,6 @@
+namespace VoW.Api.Contracts.Auth;
+
+public sealed record AuthTokenResponse(
+    string AccessToken,
+    string RefreshToken,
+    DateTimeOffset ExpiresAt);
