+++ vow-api/src/VoW.Api/Contracts/Reports/ReportSearchResponse.cs
@@ -1,0 +1,6 @@
+namespace VoW.Api.Contracts.Reports;
+
+public sealed record ReportSearchResponse(
+    int Total,
+    int Page,
+    IReadOnlyList<ReportSearchResult> Results);
