+++ vow-api/src/VoW.Api/Contracts/Reports/ReportSearchResult.cs
@@ -1,0 +1,9 @@
+namespace VoW.Api.Contracts.Reports;
+
+public sealed record ReportSearchResult(
+    int ReportId,
+    string? NpcName,
+    string ChatMessage,
+    string Status,
+    int ReportedTimes,
+    DateTime TimeSubmitted);
