+++ vow-api/src/VoW.Api/Domain/Auth/ExternalUserIdentity.cs
@@ -1,0 +1,6 @@
+namespace VoW.Api.Domain.Auth;
+
+public sealed record ExternalUserIdentity(
+    string Provider,
+    string Id,
+    string DisplayName);
