--- vow-api/src/VoW.Api/Models/AuthModels.cs
@@ -1,25 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace VoW.Api.Models;
-
-public sealed record AuthTokenResponse(
-    string AccessToken,
-    string RefreshToken,
-    DateTimeOffset ExpiresAt);
-
-public sealed record RefreshTokenRequest(
-    [Required] string RefreshToken);
-
-public sealed record AuthHandoffRequest(
-    [Required] string Code);
-
-public sealed record DiscordTokenResponse(
-    string AccessToken,
-    string TokenType,
-    int ExpiresIn,
-    string Scope);
-
-public sealed record ExternalUserIdentity(
-    string Provider,
-    string Id,
-    string DisplayName);
