+++ vow-api/src/VoW.Api/Repositories/UserRepository.cs
@@ -1,6 +1,6 @@
 using Dapper;
 using MySqlConnector;
-using VoW.Api.Models;
+using VoW.Api.Domain.Users;
 
 namespace VoW.Api.Repositories;
